+++ test/ext/runtests.jl
@@ -1,0 +1,5 @@
+
+using Ark
+using Test
+
+include("test_mooncake_interop.jl")